+++ tests/testthat/test-collocations.R
@@ -1,4 +1,5 @@
 test_that("collocationScoreQuery works", {
+  skip_if_offline()
   kco <- new("KorAPConnection", cache = TRUE, verbose = TRUE)
   df <- collocationScoreQuery(kco, "Ameisenplage", "heimgesucht", leftContextSize=0, rightContextSize=1)
   expect_gt(df$logDice, 1)
@@ -11,7 +12,8 @@ test_that("collocationScoreQuery works", {
 
 
 test_that("collocationAnalysis works and warns about missing token", {
-    kco <- new("KorAPConnection",
+  skip_if_offline()
+  kco <- new("KorAPConnection",
                accessToken = NULL,
                verbose = TRUE)
     expect_warning(
@@ -24,7 +26,8 @@ test_that("collocationAnalysis works and warns about missing token", {
           rightContextSize = 0,
           searchHitsSampleLimit = 100,
           topCollocatesLimit = 1,
-          exactFrequencies = FALSE
+          exactFrequencies = FALSE,
+          maxRecurse = 2
         ),
       "access token"
     )
@@ -33,10 +36,17 @@ test_that("collocationAnalysis works and warns about missing token", {
 })
 
 test_that("collocationAnalysis on unaccounted strings does not error out", {
+  skip_if_offline()
   kco <- new("KorAPConnection", accessToken = NULL, verbose = TRUE)
   expect_warning(
-    df <- collocationAnalysis(kco, "XXXXXXXXAmeisenplage"),
+    df <- collocationAnalysis(kco, "XXXXXXXXAmeisenplage", vc=c("corpusSigle=/WDD17/", "corpusSigle=/WUD17/"), maxRecurse = 2),
     "access token"
   )
   testthat::expect_equal(nrow(df), 0)
+})
+
+test_that("temoveWithinSpanWorks", {
+  expect_equal(
+    removeWithinSpan("contains(<base/s=s>, (machen []{0,1} aufmerksam | aufmerksam []{0,1} machen))", "base/s=s"),
+    "(machen []{0,1} aufmerksam | aufmerksam []{0,1} machen)")
 })